requirements.txt
flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
twilio==8.10.0
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
gunicorn==21.2.0